memberDTO 수정 반영
